+++ R/Utilities.R
@@ -1,4 +1,4 @@
-system.file2 = function(..., package = "GRANBase") {
+system.file2 <- function(..., package = "GRANBase") {
     ret = tryCatch(system.file(..., package = package), error = function(e) e)
 
     if(!is.null(ret) && !is(ret, "error") && nchar(ret))
@@ -15,48 +15,7 @@ system.file2 = function(..., package = "GRANBase") {
 }
 
 
-#' writeGRANLog
-#'
-#' Utility function which writes gran logs
-#' @param pkg The name of the package the log is about
-#' @param msg The log message, collapsed if length>1
-#' @param type "full", "error", "warn", or "both" indicating which
-#' log(s) the message should be written to
-#' @param logfile The location of the full log file to write/append to
-#' @param errfile the location of the error log file to write/append to
-#' @param pkglog character. The package-specific log file to write to if
-#' applicable.
-#' @note This function is not intended for direct use by the end user.
-#' @export
-writeGRANLog = function(pkg, msg, type = "full", logfile,
-                        errfile, pkglog = NULL)
-{
-    if(type == "error") {
-        targ = errfile
-        err = " ERROR "
-    } else if (type == "both") {
-        targ = c(logfile, errfile)
-        err = " ERROR "
-    } else if (type == "warn") {
-        targ = c(logfile, errfile)
-        err = " WARNING "
-    } else {
-        targ = logfile
-        err = character()
-    }
-    if(!is.null(pkg) && !is.na(pkg))
-        targ = c(targ, pkglog)
-
-    fullmsg <- paste("\n", err, "pkg:", pkg, "(", date(), ") - ",
-        paste(paste0("\t",msg), collapse="\n\t"))
-    sapply(targ, function(x) {
-      if (!file.exists(x))
-          file.create(x)
-      cat(fullmsg, append = TRUE, file = x)
-    })
-}
-
-getPkgNames = function(path)
+getPkgNames <- function(path)
 {
     path = normalizePath2(path)
     if(length(path) > 1)
@@ -68,14 +27,14 @@ getPkgNames = function(path)
 }
 
 
-getCheckoutLocs = function(codir, manifest = manifest_df(repo),
+getCheckoutLocs <- function(codir, manifest = manifest_df(repo),
     branch = manifest$branch, repo)
 {
     mapply(getPkgDir, basepath = codir, subdir = manifest$subdir,
            scm_type = manifest$type, branch = branch, name = manifest$name)
 }
 
-getMaintainers = function(codir, manifest = manifest_df(repo),
+getMaintainers <- function(codir, manifest = manifest_df(repo),
     branch = manifest$branch, repo) {
     sapply(getCheckoutLocs(codir, manifest = manifest), function(x) {
         if(!file.exists(file.path(x,"DESCRIPTION")))
@@ -89,7 +48,7 @@ getMaintainers = function(codir, manifest = manifest_df(repo),
     })
 }
 
-getCOedVersions = function(codir, manifest = manifest_df(repo),
+getCOedVersions <- function(codir, manifest = manifest_df(repo),
     branch = manifest$branch, repo) {
     locs = getCheckoutLocs(codir, manifest = manifest,
         branch = branch, repo = repo)
@@ -113,7 +72,7 @@ getCOedVersions = function(codir, manifest = manifest_df(repo),
 }
 
 
-isOkStatus = function(status= repo_results(repo)$status,
+isOkStatus <- function(status= repo_results(repo)$status,
     repo)
 {
     #status can be NA when the package isn't being built at all
@@ -122,7 +81,7 @@ isOkStatus = function(status= repo_results(repo)$status,
                       (check_note_ok(repo) & status == "check note(s)"))
 }
 
-install.packages2 = function(pkgs, repos, lib,  ..., param = SwitchrParam(),
+install.packages2 <- function(pkgs, repos, lib,  ..., param = SwitchrParam(),
     outdir = tempdir())
 {
 
@@ -136,8 +95,7 @@ install.packages2 = function(pkgs, repos, lib,  ..., param = SwitchrParam(),
     ##install.packages(pkgs, ..., keep_outputs=outdir)
     avail = available.packages(contrib.url(repos, type = "source"))
     install.packages(pkgs = pkgs, repos = repos,
-                         INSTALL_opts = sprintf("-l %s", lib), lib = lib,
-                         ..., keep_outputs=TRUE)
+                     lib = lib, ..., keep_outputs=TRUE)
     ret = sapply(pkgs, function(p)
     {
         if(! p %in% avail[,"Package"])
@@ -156,40 +114,25 @@ install.packages2 = function(pkgs, repos, lib,  ..., param = SwitchrParam(),
 }
 
 
-getBuilding = function(repo, results= repo_results(repo))
+getBuilding <- function(repo, results= repo_results(repo))
 {
     results$building & isOkStatus( repo = repo)
 }
 
-getBuildingManifest = function(repo, results = repo_results(repo),
+getBuildingManifest <- function(repo, results = repo_results(repo),
     manifest = manifest_df(repo))
 {
     manifest[getBuilding(repo, results),]
 }
 
 
-getBuildingResults = function(repo, results = repo_results(repo))
+getBuildingResults <- function(repo, results = repo_results(repo))
 {
     results[getBuilding(repo, results),]
 }
 
-builtPkgExt = function(regex = FALSE) {
-##    if(Sys.info()["sysname"] == "Darwin")
-##        ret = ".tgz"
-    ##    else if (.Platform$OS.type == "windows")
-    if (.Platform$OS.type == "windows")
-        ret = ".zip"
-    else
-        ret = ".tar.gz"
 
-    if(regex)
-        ret = gsub(".", "\\.", fixed=TRUE, ret)
-    ret
-
-}
-
-
-trim_PACKAGES = function(dir) {
+trim_PACKAGES <- function(dir) {
 
     pkgs = read.dcf(file.path(dir, "PACKAGES"))
     pkgsdf = as.data.frame(pkgs)
@@ -197,7 +140,7 @@ trim_PACKAGES = function(dir) {
         fils = pkgsdf$Files
     else {
         fils = file.path(dir, paste0(pkgsdf$Package, "_",
-                                     pkgsdf$Version, builtPkgExt()))
+                                     pkgsdf$Version, ".tar.gz"))
     }
     missing = !file.exists(fils)
     pkgsdf = pkgsdf[!missing,]
@@ -210,7 +153,7 @@ trim_PACKAGES = function(dir) {
     invisible(pkgs[missing, "Package"])
 }
 
-haveGit = function() nchar(Sys.which("git")) > 0
+haveGit <- function() nchar(Sys.which("git")) > 0
 
 substrRight <- function(x, n){
   substr(x, nchar(x)-n+1, nchar(x))
@@ -250,7 +193,7 @@ getOS <- function(){
     os <- sysinf['sysname']
     if (os == 'Darwin')
       os <- "osx"
-  } else { ## mystery machine
+  } else {
     os <- .Platform$OS.type
     if (grepl("^darwin", R.version$os))
       os <- "osx"