+++ inst/GRAN/R/myGRAN.R
@@ -2,10 +2,10 @@ myGRANenv = new.env()
 ##' defaultGRAN
 ##'
 ##' Get the GRAN repository object for the associated repo
-##' 
+##'
 ##' @return The GRANRepository object encapsulated by this package
 ##' @export
-defaultGRAN = function() {
+defaultGRAN <- function() {
     if(is.null(myGRANenv$repo))
         myGRANenv$repo = loadRepo(system.file("myrepo.R", package = pkgname))
     myGRANenv$repo
@@ -17,4 +17,4 @@ defaultGRAN = function() {
 ##' @return A character value with the URL for the default GRAN repository in
 ##' this package
 ##' @export
-defaultGRANURL = function() repo_url(defaultGRAN())
+defaultGRANURL <- function() repo_url(defaultGRAN())